Bug fixes: - Uses padding value type that is the same as the atten_mask, which may be different from float32 (e.g., on TPU). - Position ("segment_pos") should be created (when None) as int rather than float tensor. - Avoids get_decode_state on 'ngrammer_prefix_segment_pos' when segment_pos is None, as the state would be neither needed nor available.
PiperOrigin-RevId: 610745672

+++ praxis/layers/embedding_softmax.py
@@ -1321,7 +1321,7 @@ class TrainablePositionalEmbedding(PositionalEmbedding):
     ap = self.activation_split_dims_mapping
     if position is None:
       assert seq_length is not None
-      position = jnp.arange(seq_length, dtype=jnp.float32)[jnp.newaxis, :]
+      position = jnp.arange(seq_length, dtype=jnp.int32)[jnp.newaxis, :]
     if seq_length is None:
       assert position is not None
       assert position.ndim == 2
